airline-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
pydantic-settings
python-jose[cryptography]
bcrypt>=4.0.0
python-multipart
gunicorn
